llm_agents/advanced_agent.py
from typing import List, Dict, Any, Optional, Union
from pydantic import BaseModel, Field
import datetime
import json
import asyncio
from concurrent.futures import ThreadPoolExecutor
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity
from sentence_transformers import SentenceTransformer

class AgentMemory(BaseModel):
    short_term: List[Dict[str, Any]] = Field(default_factory=list)
    long_term: List[Dict[str, Any]] = Field(default_factory=list)
    embeddings: Dict[str, np.ndarray] = Field(default_factory=dict)
    max_short_term: int = 20
    max_long_term: int = 1000
    embedding_model: Any = None

    def __init__(self, **data):
        super().__init__(**data)
        self.embedding_model = SentenceTransformer('all-MiniLM-L6-v2')

    def add_short_term(self, thought: str, action: str, result: str, metadata: Dict[str, Any] = None):
        memory_entry = {
            "thought": thought,
            "action": action,
            "result": result,
            "timestamp": datetime.datetime.now(),
            "metadata": metadata or {}
        }
        self.short_term.append(memory_entry)
        if len(self.short_term) > self.max_short_term:
            self.short_term.pop(0)

    def add_long_term(self, key: str, value: Any, metadata: Dict[str, Any] = None):
        memory_entry = {
            "key": key,
            "value": value,
            "timestamp": datetime.datetime.now(),
            "metadata": metadata or {}
        }
        self.long_term.append(memory_entry)
        
        # Store embedding for semantic search
        embedding = self.embedding_model.encode(str(value))
        self.embeddings[key] = embedding
        
        if len(self.long_term) > self.max_long_term:
            oldest_entry = self.long_term.pop(0)
            if oldest_entry["key"] in self.embeddings:
                del self.embeddings[oldest_entry["key"]]

    def get_relevant_memory(self, query: str, top_k: int = 5) -> List[Dict[str, Any]]:
        query_embedding = self.embedding_model.encode(query)
        similarities = {}
        
        for key, embedding in self.embeddings.items():
            similarity = cosine_similarity([query_embedding], [embedding])[0][0]
            similarities[key] = similarity
        
        sorted_keys = sorted(similarities.items(), key=lambda x: x[1], reverse=True)[:top_k]
        return [next(item for item in self.long_term if item["key"] == key) for key, _ in sorted_keys]

class AgentGoal(BaseModel):
    main_goal: str
    sub_goals: List[str] = Field(default_factory=list)
    status: str = "in_progress"
    progress: float = 0.0
    created_at: datetime.datetime = Field(default_factory=datetime.datetime.now)
    updated_at: datetime.datetime = Field(default_factory=datetime.datetime.now)

    def update_progress(self, progress: float):
        self.progress = progress
        self.updated_at = datetime.datetime.now()

    def add_sub_goal(self, sub_goal: str):
        self.sub_goals.append(sub_goal)
        self.updated_at = datetime.datetime.now()

class AdvancedAgent(BaseModel):
    llm: Any
    tools: List[Any]
    memory: AgentMemory = Field(default_factory=AgentMemory)
    current_goal: Optional[AgentGoal] = None
    confidence_threshold: float = 0.7
    max_loops: int = 20
    max_parallel_tools: int = 3
    tool_stats: Dict[str, Dict[str, int]] = Field(default_factory=dict)
    learning_rate: float = 0.1

    def __init__(self, **data):
        super().__init__(**data)
        self.tool_stats = {tool.name: {"success": 0, "failure": 0} for tool in self.tools}

    async def run(self, goal: str) -> str:
        self.current_goal = AgentGoal(main_goal=goal)
        self._decompose_goal(goal)
        
        previous_responses = []
        num_loops = 0
        
        while num_loops < self.max_loops and self.current_goal.status != "completed":
            num_loops += 1
            context = self._build_context(previous_responses)
            
            # Generate next action with confidence
            action, confidence = await self._generate_action(context)
            
            if confidence < self.confidence_threshold:
                # Implement fallback strategy
                action = await self._generate_fallback_action(context)
            
            # Execute action
            result = await self._execute_action(action)
            
            # Update memory and goal progress
            self._update_memory(action, result)
            self._update_goal_progress(result)
            
            previous_responses.append((action, result))
            
            # Check if goal is completed
            if self._is_goal_completed(result):
                self.current_goal.status = "completed"
                break
        
        return self._generate_final_answer(previous_responses)

    async def _generate_action(self, context: str) -> tuple:
        prompt = self._build_action_prompt(context)
        response = await self.llm.generate(prompt)
        action = self._parse_action(response)
        confidence = self._calculate_confidence(response)
        return action, confidence

    async def _execute_action(self, action: Dict[str, Any]) -> str:
        if action["type"] == "tool":
            return await self._execute_tool(action)
        elif action["type"] == "sub_goal":
            return await self._execute_sub_goal(action)
        else:
            raise ValueError(f"Unknown action type: {action['type']}")

    async def _execute_tool(self, action: Dict[str, Any]) -> str:
        tool_name = action["tool"]
        tool_input = action["input"]
        
        try:
            # Execute tool with timeout
            result = await asyncio.wait_for(
                self._execute_tool_with_timeout(tool_name, tool_input),
                timeout=30.0
            )
            self.tool_stats[tool_name]["success"] += 1
            return result
        except Exception as e:
            self.tool_stats[tool_name]["failure"] += 1
            return f"Error executing tool {tool_name}: {str(e)}"

    async def _execute_tool_with_timeout(self, tool_name: str, tool_input: str) -> str:
        tool = next(t for t in self.tools if t.name == tool_name)
        return await tool.use(tool_input)

    def _update_memory(self, action: Dict[str, Any], result: str):
        self.memory.add_short_term(
            thought=action.get("thought", ""),
            action=f"{action['type']}: {action.get('tool', '')}",
            result=result,
            metadata={"confidence": action.get("confidence", 0.0)}
        )

    def _update_goal_progress(self, result: str):
        # Update goal progress based on result
        progress = self._calculate_progress(result)
        self.current_goal.update_progress(progress)

    def _calculate_progress(self, result: str) -> float:
        # Implement progress calculation logic
        return min(1.0, self.current_goal.progress + 0.1)

    def _is_goal_completed(self, result: str) -> bool:
        # Implement goal completion check
        return self.current_goal.progress >= 1.0

    def _generate_final_answer(self, previous_responses: List[tuple]) -> str:
        # Generate final answer based on all responses
        return "Final answer based on all interactions"

    def _build_context(self, previous_responses: List[tuple]) -> str:
        # Build context from memory and previous responses
        return "Context built from memory and responses"

    def _decompose_goal(self, goal: str):
        # Decompose main goal into sub-goals
        self.current_goal.add_sub_goal("First sub-goal")
        self.current_goal.add_sub_goal("Second sub-goal")

    def _calculate_confidence(self, response: str) -> float:
        # Implement confidence calculation
        return 0.9

    def _parse_action(self, response: str) -> Dict[str, Any]:
        # Parse LLM response into action
        return {"type": "tool", "tool": "example", "input": "example"}

    async def _generate_fallback_action(self, context: str) -> Dict[str, Any]:
        # Generate fallback action when confidence is low
        return {"type": "tool", "tool": "fallback", "input": "fallback"} 