requirements.txt
openai>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
sentence-transformers>=2.2.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
langchain>=0.1.0
tiktoken>=0.4.0
tenacity>=8.2.0
asyncio>=3.4.3
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0